passes tests
Added the Mythical65c24T8 as a third type of 65xx CPU.  It is (so far) the WDC65C02 with two extra opcodes: 0F=CPU to fill A with capabilities and 1F=A24 as a prefix code to switch the next opcode into 24-bit address mode.

Without the prefix code, the CPU acts identically to the WDC65C02 (no Rockwell opcodes).